requirements.txt
flask
supabase==2.9.0
python-dotenv
